+++ src/features/animations/components/AnimationCard/style.ts
