+++ src/features/animations/components/AnimationCarousel/style.ts
