+++ src/features/animations/components/AnimationSlide/style.ts
